Drop premult-approach
Instead of using a premult function for creating a substituted version
of the salt for a second inclusion in the hash, separate all components
to be hashed by a byte string which gets longer for every iteration.

This should protect against symbolic optimization attacks even better,
and is much easier to achieve because the contents of the separator
string do not really matter - only that its size is different for every
iteration.