Omit padding for the first iteration
The purpose of int2pad() is to avoid feeding the same components at the
same offsets into the hash function when iterating.

There is no need to always insert padding, just that the amount of
padding is different for all iterations.

Therefore the first iteration does not use any padding at all now, then
one byte of padding for the second iteration, etc.

This makes the simple case that only a single iteration is used
especially easy to implement, as no padding is required at all in this
case.