Drop password-stretching
It maked the algorithm too complicated and helps only in scenarios where
the attacker has too limited resources to run brute-force attacks on all
passwords.

This is a best-case scenario (for the victim) which cannot be generally
by assumed.

Password-stretching is also ineffective of only a single password is
being attacked, because then nothing in the world can make a password
secure enough which is way too short.

Only the iteration count can combat this to a minor degree, but this
works fine already without any requirement for password stretching.