Ignore TMPDIR
Instead of relying on the TMPDIR environment variable for determining
the location of a directory for temporary files containing sensitive
information, provide a -t option for explicitly specifying the directory
or use a hard-coded default (/tmp) as a fallback.

This protects against an incorrectly set $TMPDIR which does not refer to
a secure location.