Renamed the script again to include "o" for "open"
This shall be a reminder that the script does not rely on the SSL
protocol but rather on the OpenSSL implementation of the protocol, which
also includes cryptographic primitives, because only those primitives
are actually used by the script.

The script does not have anything to do with the "failure by design" SSL
protocol itself!