Make requirement for iconv optional
iconv is required only if the current locale is not already using UTF-8
or or if the UTF-8 password contains decomposed characters which need to
be re-composed.

iconv is never required if the password consists only of
ASCII-characters, or if the user does not care about normalization
(meaning different UTF-8 representations of the same password will be
considered to be different passwords).

Avoiding to include iconv might be especially desirable for
initramfs-based preboot-environments, because iconv requires large
tables to be installed which makes the initramfs image much larger than
without iconv support.