Use NFKC rather than NFC
The advantage of using NFKC is that there is a tool available for
converting any string into NFKC, and iconv is no longer required.

Another advantage is that visually identically symbols represented by
different code points will be mapped into the same code point, using the
same mapping algorithms as for internationalized domain names.

This is especially important when printing passwords on paper, because
otherwise it would not be known which code points to actually use to
represent those characters.

The option to bypass the normalization stop has been renamed to "-N" for
"no normalization".