Support script file extensions and non-executables
Allow both scripts to be renamed to an arbitrary extension such as *.sh,
without a need to modify the contents of the scripts for that.

Also, allow both scripts to be plain text files without requirement for
executable permissions, which are invoked directly by passing the script
name as an argument to the "sh" command.

Together, this will allow both scripts to be invoked directly from a
FAT-filesystem which does not support the concept of an "executable"-
permission (depending on the mount options).

Supporting to rename the files to an arbitrary file extension (such as
".sh") allows to register a custom interpreter for Bourne Shell scripts
with the Microsoft Windows registry, which does not support the UNIX
"shebang"-mechanism at all.