Support the script without executable permissions
The script can be run directly by some shell, rather than being
looked up via the shebang mechansim.

This allows the script to be run, for instance, from a file system which
does not provide executable permissions.

Add the -s option for specifying the shell to be used for invoking
helper shell scripts.

This is better than relying on $SHELL, because the user might use a
shell which is not compatible with the POSIX shell.

"sh" will used as a default for -s. But "sh" is not always available,
for instance when running from a BusyBox system where no symlinks have
been set up for the utilities, and every utility including the shell
needs to be invoked as an argument of the "busybox" command. Option -s
supports such installations.